+++ packages/blocks-core/js/libs/wordpress/column/index.js
@@ -50,6 +50,12 @@ export const Column: BlockType = {
 		'core/heading-4': sharedInnerBlocks['core/heading-4'],
 		'core/heading-5': sharedInnerBlocks['core/heading-5'],
 		'core/heading-6': sharedInnerBlocks['core/heading-6'],
+		'elements/bold': sharedInnerBlocks['elements/bold'],
+		'elements/italic': sharedInnerBlocks['elements/italic'],
+		'elements/kbd': sharedInnerBlocks['elements/kbd'],
+		'elements/code': sharedInnerBlocks['elements/code'],
+		'elements/span': sharedInnerBlocks['elements/span'],
+		'elements/mark': sharedInnerBlocks['elements/mark'],
 	},
 	edit: (props) => {
 		return <SharedBlockExtension {...props} />;