+++ packages/blocks-core/CHANGELOG.md
@@ -24,6 +24,7 @@
 - `Paragraph` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
 - `Button` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
 - `Buttons` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
+- `Column` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
 - `Columns` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
 - `Accordion Heading` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
 
@@ -51,6 +52,7 @@
 - Added E2E tests to check functionality `Paragraph` new inner block.
 - Added E2E tests to check functionality `Button` new inner block.
 - Added E2E tests to check functionality `Buttons` new inner block.
+- Added E2E tests to check functionality `Column` new inner block.
 - Added E2E tests to check functionality `Columns` new inner block.
 - Added E2E tests to check functionality `Accordion Heading` new inner block.